refactor: 불필요한 component 삭제, 북마크 api 통합
