--- src/components/statics_prolog/Guide2.vue
@@ -1,106 +1,0 @@
-<script setup></script>
-
-<template>
-  <div class="section">
-    <h2>2. 사전 준비 단계</h2>
-    <p></p>
-    <div class="subsection">
-      <h3>🔹 청약통장 가입 여부 확인</h3>
-      <ul>
-        <li>필수 조건 중 하나. 없으면 청약 불가</li>
-        <li>종류: 주택청약종합저축, 청년우대형 등</li>
-        <li>
-          <strong>가입 기간</strong>과 <strong>납입 횟수</strong>에 따라 가점
-          결정
-        </li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 본인 조건 파악</h3>
-      <ul>
-        <li><strong>무주택 여부</strong>: 무주택자 우선 대상</li>
-        <li><strong>가구 구성</strong>: 세대주, 부양가족 수 등도 가점 요소</li>
-        <li><strong>소득 및 자산 기준</strong>: 특별공급은 소득 제한 있음</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 자주 활용되는 사이트 회원가입</h3>
-      <ul>
-        <li>
-          <a href="https://www.applyhome.co.kr/" target="_blank">청약홈</a>:
-          청약 신청 및 결과 조회
-        </li>
-        <li>LH, SH 등 지역공사 사이트: 공공임대/분양 정보 확인용</li>
-        <li>공동인증서, 휴대폰 인증 등 <strong>본인 인증 준비</strong> 필요</li>
-      </ul>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.guide {
-  line-height: 1.8;
-  font-size: 16px;
-  color: #222;
-}
-
-h2 {
-  font-size: 28px;
-  margin-bottom: 24px;
-  border-bottom: 2px solid #ccc;
-  padding-bottom: 10px;
-}
-
-h3 {
-  font-size: 22px;
-  margin-top: 40px;
-  margin-bottom: 16px;
-  color: #005b99;
-  border-left: 5px solid #005b99;
-  padding-left: 12px;
-}
-
-h4 {
-  font-size: 18px;
-  margin-top: 24px;
-  margin-bottom: 10px;
-  color: #333;
-  font-weight: bold;
-}
-
-ul,
-ol {
-  padding-left: 20px;
-  margin-bottom: 20px;
-}
-
-ul li,
-ol li {
-  margin-bottom: 8px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin: 20px 0;
-  font-size: 15px;
-}
-
-th,
-td {
-  border: 1px solid #ccc;
-  padding: 10px;
-  text-align: left;
-}
-
-blockquote {
-  background: #f9f9f9;
-  padding: 14px;
-  margin: 24px 0;
-  border-left: 5px solid #00aaff;
-  font-style: italic;
-  color: #333;
-}
-</style>
