--- src/components/statics_prolog/Guide3.vue
@@ -1,99 +1,0 @@
-<script setup></script>
-
-<template>
-  <div class="section">
-    <h2>3. 청약 정보 탐색</h2>
-
-    <div class="subsection">
-      <h3>🔹 청약 일정 확인</h3>
-      <ul>
-        <li>청약홈, LH, SH 등에서 <strong>청약 공고 일정</strong> 수시 확인</li>
-        <li>접수 기간이 짧으므로 <strong>사전 확인 필수</strong></li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 관심 단지 선별</h3>
-      <ul>
-        <li>입지: 교통, 학교, 편의시설 등</li>
-        <li>분양가: 예산에 맞는 수준인지 확인</li>
-        <li>면적: 가족 수에 적합한 평형 선택</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 모집공고 꼼꼼히 읽기</h3>
-      <ul>
-        <li>공급 유형: 민영/공공, 특별/일반공급 여부</li>
-        <li>자격 요건: 연령, 무주택 여부, 가점 기준 등</li>
-        <li>준비 서류 및 접수 일정 확인</li>
-      </ul>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.guide {
-  line-height: 1.8;
-  font-size: 16px;
-  color: #222;
-}
-
-h2 {
-  font-size: 28px;
-  margin-bottom: 24px;
-  border-bottom: 2px solid #ccc;
-  padding-bottom: 10px;
-}
-
-h3 {
-  font-size: 22px;
-  margin-top: 40px;
-  margin-bottom: 16px;
-  color: #005b99;
-  border-left: 5px solid #005b99;
-  padding-left: 12px;
-}
-
-h4 {
-  font-size: 18px;
-  margin-top: 24px;
-  margin-bottom: 10px;
-  color: #333;
-  font-weight: bold;
-}
-
-ul,
-ol {
-  padding-left: 20px;
-  margin-bottom: 20px;
-}
-
-ul li,
-ol li {
-  margin-bottom: 8px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin: 20px 0;
-  font-size: 15px;
-}
-
-th,
-td {
-  border: 1px solid #ccc;
-  padding: 10px;
-  text-align: left;
-}
-
-blockquote {
-  background: #f9f9f9;
-  padding: 14px;
-  margin: 24px 0;
-  border-left: 5px solid #00aaff;
-  font-style: italic;
-  color: #333;
-}
-</style>
