--- src/components/statics_prolog/Guide4.vue
@@ -1,101 +1,0 @@
-<script setup></script>
-
-<template>
-  <div class="section">
-    <h2>4. 청약 자격 판단 및 유형 선택</h2>
-
-    <div class="subsection">
-      <h3>🔹 특별공급 해당 여부 확인</h3>
-      <ul>
-        <li>신혼부부, 다자녀, 생애최초, 노부모 부양 등</li>
-        <li>각 유형마다 소득·자산·무주택 요건 다름</li>
-        <li>해당 시 일반공급보다 <strong>우선 청약 가능</strong></li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 일반공급 자격 확인</h3>
-      <ul>
-        <li>
-          <strong>가점제</strong>: 무주택 기간, 부양가족 수, 청약통장 가입기간
-        </li>
-        <li><strong>추첨제</strong>: 가점과 관계없이 무작위 추첨</li>
-        <li>단지마다 가점제/추첨제 <strong>비율 다름</strong></li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 청약 신청 가능 회차 확인</h3>
-      <ul>
-        <li>동일 세대 내 <strong>1인만 신청 가능</strong></li>
-        <li><strong>1순위 자격</strong> 여부 확인: 지역별 기준 상이</li>
-      </ul>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.guide {
-  line-height: 1.8;
-  font-size: 16px;
-  color: #222;
-}
-
-h2 {
-  font-size: 28px;
-  margin-bottom: 24px;
-  border-bottom: 2px solid #ccc;
-  padding-bottom: 10px;
-}
-
-h3 {
-  font-size: 22px;
-  margin-top: 40px;
-  margin-bottom: 16px;
-  color: #005b99;
-  border-left: 5px solid #005b99;
-  padding-left: 12px;
-}
-
-h4 {
-  font-size: 18px;
-  margin-top: 24px;
-  margin-bottom: 10px;
-  color: #333;
-  font-weight: bold;
-}
-
-ul,
-ol {
-  padding-left: 20px;
-  margin-bottom: 20px;
-}
-
-ul li,
-ol li {
-  margin-bottom: 8px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin: 20px 0;
-  font-size: 15px;
-}
-
-th,
-td {
-  border: 1px solid #ccc;
-  padding: 10px;
-  text-align: left;
-}
-
-blockquote {
-  background: #f9f9f9;
-  padding: 14px;
-  margin: 24px 0;
-  border-left: 5px solid #00aaff;
-  font-style: italic;
-  color: #333;
-}
-</style>
