--- src/components/statics_prolog/Guide6.vue
@@ -1,104 +1,0 @@
-<script setup></script>
-
-<template>
-  <div class="section">
-    <h2>6. 청약 후 절차</h2>
-
-    <div class="subsection">
-      <h3>🔹 청약 결과 확인</h3>
-      <ul>
-        <li>청약홈 > 당첨조회 메뉴</li>
-        <li>발표일에 결과 확인 가능</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 당첨 시</h3>
-      <ul>
-        <li>서류 제출 기간 및 방법 확인</li>
-        <li>분양계약 일정 확인 및 준비 (계약금 등)</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 낙첨 시</h3>
-      <ul>
-        <li>가점 유지 여부 확인</li>
-        <li>다음 청약 도전을 위한 조건 점검</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 서류 제출 누락 주의</h3>
-      <ul>
-        <li>당첨 후 서류 미제출 시 부적격 처리 가능</li>
-      </ul>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.guide {
-  line-height: 1.8;
-  font-size: 16px;
-  color: #222;
-}
-
-h2 {
-  font-size: 28px;
-  margin-bottom: 24px;
-  border-bottom: 2px solid #ccc;
-  padding-bottom: 10px;
-}
-
-h3 {
-  font-size: 22px;
-  margin-top: 40px;
-  margin-bottom: 16px;
-  color: #005b99;
-  border-left: 5px solid #005b99;
-  padding-left: 12px;
-}
-
-h4 {
-  font-size: 18px;
-  margin-top: 24px;
-  margin-bottom: 10px;
-  color: #333;
-  font-weight: bold;
-}
-
-ul,
-ol {
-  padding-left: 20px;
-  margin-bottom: 20px;
-}
-
-ul li,
-ol li {
-  margin-bottom: 8px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin: 20px 0;
-  font-size: 15px;
-}
-
-th,
-td {
-  border: 1px solid #ccc;
-  padding: 10px;
-  text-align: left;
-}
-
-blockquote {
-  background: #f9f9f9;
-  padding: 14px;
-  margin: 24px 0;
-  border-left: 5px solid #00aaff;
-  font-style: italic;
-  color: #333;
-}
-</style>
