--- src/components/statics_prolog/Guide7.vue
@@ -1,109 +1,0 @@
-<script setup></script>
-
-<template>
-  <div class="section">
-    <h2>7. 유의사항</h2>
-
-    <div class="subsection">
-      <h3>🔹 부적격 당첨 시 청약 제한</h3>
-      <ul>
-        <li>최대 3년간 청약 신청 불가</li>
-        <li>서류 미제출, 자격 미달 등 포함</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 허위 정보 입력 금지</h3>
-      <ul>
-        <li>위장전입, 부양가족 조작 등 적발 시 불이익</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 청약 횟수 및 순위 소진 주의</h3>
-      <ul>
-        <li>1순위 소진 시, 다음 청약부터 2순위로 불리해짐</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 공고문은 반드시 직접 확인</h3>
-      <ul>
-        <li>공급유형, 우선순위 조건, 제출서류 등 꼼꼼히 확인</li>
-      </ul>
-    </div>
-
-    <div class="subsection">
-      <h3>🔹 청약금 납입 여부 확인</h3>
-      <ul>
-        <li>일부 유형은 청약금 납부 필수, 미납 시 무효 처리</li>
-      </ul>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.guide {
-  line-height: 1.8;
-  font-size: 16px;
-  color: #222;
-}
-
-h2 {
-  font-size: 28px;
-  margin-bottom: 24px;
-  border-bottom: 2px solid #ccc;
-  padding-bottom: 10px;
-}
-
-h3 {
-  font-size: 22px;
-  margin-top: 40px;
-  margin-bottom: 16px;
-  color: #005b99;
-  border-left: 5px solid #005b99;
-  padding-left: 12px;
-}
-
-h4 {
-  font-size: 18px;
-  margin-top: 24px;
-  margin-bottom: 10px;
-  color: #333;
-  font-weight: bold;
-}
-
-ul,
-ol {
-  padding-left: 20px;
-  margin-bottom: 20px;
-}
-
-ul li,
-ol li {
-  margin-bottom: 8px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin: 20px 0;
-  font-size: 15px;
-}
-
-th,
-td {
-  border: 1px solid #ccc;
-  padding: 10px;
-  text-align: left;
-}
-
-blockquote {
-  background: #f9f9f9;
-  padding: 14px;
-  margin: 24px 0;
-  border-left: 5px solid #00aaff;
-  font-style: italic;
-  color: #333;
-}
-</style>
