--- src/components/statics_prolog/Guide8.vue
@@ -1,246 +1,0 @@
-<script setup></script>
-
-<template>
-  <div class="section">
-    <h2>8. 자주묻는 질문 리스트 (FAQ)</h2>
-
-    <div class="faq-item">
-      <h3>1. Q. 청약 신청은 어디서 하나요?</h3>
-      <p>
-        A. 대부분의 청약은
-        <a
-          href="https://www.applyhome.co.kr/"
-          target="_blank"
-          rel="noopener noreferrer"
-          >청약홈</a
-        >
-        에서 온라인으로 신청할 수 있습니다. 공공임대, 도시개발 관련 청약은 LH,
-        SH 등 해당 기관 사이트에서 진행됩니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>2. Q. 청약통장이 꼭 있어야 하나요?</h3>
-      <p>
-        A. 네. 대부분의 일반/특별공급 청약은
-        <strong>주택청약종합저축</strong> 등
-        <strong>청약통장이 필수</strong>입니다. 단, 일부 임대주택의 경우
-        청약통장이 없어도 가능합니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>3. Q. 무주택이어야만 청약이 가능한가요?</h3>
-      <p>
-        A. 무주택자가 우선되지만, 일부 유형(추첨제, 일부 임대청약 등)은
-        유주택자도 가능할 수 있습니다. 단, 당첨 가능성이 매우 낮습니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>4. Q. 청약 가점은 어떻게 계산되나요?</h3>
-      <p>
-        A. 가점제는
-        <strong>무주택 기간 + 부양가족 수 + 청약통장 가입기간</strong>으로 총
-        84점 만점입니다.<br />
-        청약홈에서 자동 계산기 서비스를 이용할 수 있어요.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>5. Q. 청약통장에 돈은 얼마나 넣어야 하나요?</h3>
-      <p>
-        A. 민영주택의 경우 면적에 따라
-        <strong>최대 1,500만 원까지 인정</strong>됩니다. 다달이 10만 원 이상
-        넣는 것이 일반적이며, 지역에 따라 최소 납입금도 다릅니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>6. Q. 1순위 조건은 어떻게 되나요?</h3>
-      <p>
-        A. 기본적으로
-        <strong
-          >청약통장 가입기간 2년 이상 + 지역·주택유형별 납입 조건 충족 + 무주택
-          요건</strong
-        >을 만족해야 1순위가 됩니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>7. Q. 1인 가구도 청약이 가능한가요?</h3>
-      <p>
-        A. 가능합니다. 다만 가점에서 <strong>부양가족 수가 적어 불리</strong>할
-        수 있습니다. 생애최초나 추첨제 청약을 노리는 것이 좋습니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>8. Q. 당첨되면 자동으로 계약이 되나요?</h3>
-      <p>
-        A. 아닙니다. <strong>당첨 후 별도로 분양계약을 체결해야</strong> 하며,
-        일정 내 서류 제출과 계약금 납입도 필요합니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>9. Q. 청약을 여러 곳 동시에 신청할 수 있나요?</h3>
-      <p>
-        A. <strong>동일 기간 내 1개 단지에만 신청 가능</strong>합니다. 여러 군데
-        신청하면 모두 무효 처리됩니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>10. Q. 청약 당첨되면 청약통장은 없어지나요?</h3>
-      <p>
-        A. 일반적으로
-        <strong
-          >당첨되어 계약까지 진행하면 통장이 효력을 잃고 재사용 불가</strong
-        >합니다. 하지만 낙첨되면 계속 사용 가능합니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>11. Q. 가점이 낮으면 청약해도 의미 없나요?</h3>
-      <p>
-        A. 아닙니다. <strong>가점제 외에도 추첨제 물량이 존재</strong>합니다.
-        특히 전용 85㎡ 초과 주택은 추첨제 비율이 높은 편입니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>12. Q. 신혼부부 특별공급은 언제까지 신혼으로 인정되나요?</h3>
-      <p>
-        A. <strong>혼인신고 후 7년 이내</strong>면 신혼부부 특별공급 신청이
-        가능합니다. 임신 중이거나 아이가 있는 예비부부도 조건을 만족하면
-        가능해요.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>13. Q. 위장전입이나 허위서류 제출하면 어떻게 되나요?</h3>
-      <p>
-        A. <strong>당첨 취소 및 향후 최대 10년간 청약 제한</strong> 등의 강한
-        제재를 받습니다. 실거주 조건도 반드시 확인하세요.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>14. Q. 당첨되고 계약 안 하면 불이익 있나요?</h3>
-      <p>
-        A. <strong>있습니다.</strong> 일반적으로
-        <strong>1년간 청약 자격이 제한</strong>되며, 무순위 청약은 2년 제한을
-        받을 수 있습니다.
-      </p>
-    </div>
-
-    <hr />
-
-    <div class="faq-item">
-      <h3>15. Q. 청약 결과는 어디서 확인하나요?</h3>
-      <p>
-        A. 청약홈 &gt;
-        <a
-          href="https://www.applyhome.co.kr/co/COA120M01.do"
-          target="_blank"
-          rel="noopener noreferrer"
-          >당첨자 발표조회</a
-        >
-        메뉴에서 발표일에 확인 가능합니다. 문자로도 안내가 오기도 합니다.
-      </p>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.guide {
-  line-height: 1.8;
-  font-size: 16px;
-  color: #222;
-}
-
-h2 {
-  font-size: 28px;
-  margin-bottom: 24px;
-  border-bottom: 2px solid #ccc;
-  padding-bottom: 10px;
-}
-
-h3 {
-  font-size: 22px;
-  margin-top: 40px;
-  margin-bottom: 16px;
-  color: #005b99;
-  border-left: 5px solid #005b99;
-  padding-left: 12px;
-}
-
-h4 {
-  font-size: 18px;
-  margin-top: 24px;
-  margin-bottom: 10px;
-  color: #333;
-  font-weight: bold;
-}
-
-ul,
-ol {
-  padding-left: 20px;
-  margin-bottom: 20px;
-}
-
-ul li,
-ol li {
-  margin-bottom: 8px;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin: 20px 0;
-  font-size: 15px;
-}
-
-th,
-td {
-  border: 1px solid #ccc;
-  padding: 10px;
-  text-align: left;
-}
-
-blockquote {
-  background: #f9f9f9;
-  padding: 14px;
-  margin: 24px 0;
-  border-left: 5px solid #00aaff;
-  font-style: italic;
-  color: #333;
-}
-</style>
