+++ src/config/index.js
@@ -8,20 +8,10 @@ export default {
       url: '/calendar',
       icon: 'fa-solid fa-paste',
     },
-    // {
-    //   title: '마이페이지',
-    //   url: '/mypage',
-    //   icon: 'fa-solid fa-plane-departure',
-    // },
     {
       title: '지도',
       url: '/map',
       icon: 'fa-regular fa-images',
-    },
-    {
-      title: 'FAQ',
-      url: '/prolog',
-      icon: 'fa-solid fa-check',
     },
   ],
   accoutMenus: {
